Added icon + separated 3DTilesUtils in 2 files
